Controlling the useEffect() behavior in functional components useEffect() can have combined effect for componentDidMount() & componentDidUpdate() as it runs for every render cycle. We can use as many useEffect() calls as required.
In Cockpit.js,
Suppose we have a scenario where we want useEffect() to execute only during component creation(say for http request), & only once(not for update cycle).
We can control the execution of useEffect() as shown below.
The example setTimeout() mimics http request for saving data to cloud, which needs to be executed when persons array changes.
    useEffect(() => {
      console.log('[Cockpit.js] useEffect');
      // HTTP requests...

      setTimeout(() => {
          alert('Saved Data to cloud');
      }, 1000);
    }, [props.persons])
This executes only when persons array changes or data in array changes.

case2:- If we want to execute this useEffect() only once during component creation, we can pass an empty array as argument parameter
    useEffect(() => {
      console.log('[Cockpit.js] useEffect');
      // HTTP requests...

      setTimeout(() => {
          alert('Saved Data to cloud');
      }, 1000);
    }, []);
This tells react that this useEffect() call does not have any dependency on parameters(in state), so not necessary to re-render at any time. Since, there are no dependencies, the dependencies can never change.

case3:- We can pass mutliple elements in the array as dependencies similar to single dependency in case1.